+++ android/features/android-feature-navigation/build.gradle
