+++ android/features/android-feature-navigation/consumer-rules.pro
