+++ android/features/android-feature-navigation/src/main/AndroidManifest.xml
