+++ android/features/android-feature-navigation/src/main/java/io/matthewnelson/android_feature_navigation/requests/PopBackStack.kt
@@ -13,7 +13,7 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.concept_android_navigation.requests
+package io.matthewnelson.android_feature_navigation.requests
 
 import androidx.annotation.IdRes
 import androidx.navigation.NavController
