+++ android/features/android-feature-navigation/src/main/res/anim/slide_in_top.xml
