+++ android/features/android-feature-navigation/src/main/res/anim/slide_out_right.xml
