+++ android/features/android-feature-views/build.gradle
@@ -33,7 +33,7 @@ android {
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
-    api project(path: ':android:concepts:concept-android-navigation')
+    api project(path: ':android:features:android-feature-navigation')
     api project(path: ':android:features:android-feature-toast-utils')
     api project(path: ':kotlin:concepts:concept-coroutines')
     api project(path: ':kotlin:concepts:concept-views')
